% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dep.R
\name{dep_graph}
\alias{dep_graph}
\title{Dependency graph}
\usage{
dep_graph(x, ...)
}
\arguments{
\item{x}{A list with elements \code{refId}, \code{name}, \code{description}
and optional elements executables (nested packages) and references
(dependencies) with \code{from} and \code{to} elements.}

\item{...}{Additional graphviz attributes}
}
\description{
Returns a graphviz dot source for dependency graph
}
\examples{
test1 <-
  list(
    refId = "01", name = "Package 01", executables = list(
      list(refId = "0101", name = "Package 0101"),
      list(refId = "0102", name = "Package 0102")
    ),
    references = list(
      list(from = "0101", to = "0102")
    )
  )

dot_source <- dep_graph(test1)
}

