% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dep.R
\docType{package}
\name{depmodelr}
\alias{depmodelr}
\alias{depmodelr-package}
\title{depmodelr}
\description{
Create module dependency graphs
}
\examples{
\dontrun{
# read yaml file
library(yaml)
dep1 <- yaml.load_file(system.file("example.yml", package = "depmodelr") )

# create dependency graph
g1 <- dep_graph(dep1)

# display graph
library(DiagrammeR)
grViz(g1)

}
}
\seealso{
\code{\link{dep_graph}}
}

